Added new project Hook
